+++ spec/components/elements/molecules/KeyPointTransform.spec.ts
@@ -20,7 +20,7 @@ Copyright (C) 2021, Tomoya Komiyama.
 import { mount } from '@vue/test-utils'
 import Target from '/@/components/elements/molecules/KeyPointTransform.vue'
 import { getKeyframeBone, getKeyframePoint } from '/@/models/keyframe'
-import { getKeyframeBoneDefaultPropsMap } from '/@/utils/keyframes'
+import { getKeyframeDefaultPropsMap } from '/@/utils/keyframes'
 
 describe('src/components/elements/molecules/KeyPointTransform.vue', () => {
   describe('snapshot', () => {
@@ -53,7 +53,7 @@ describe('src/components/elements/molecules/KeyPointTransform.vue', () => {
             scaleX: getKeyframePoint(),
           }),
           expanded: true,
-          sameRangeWidth: getKeyframeBoneDefaultPropsMap(() => 1),
+          sameRangeWidth: getKeyframeDefaultPropsMap(() => 1),
         },
       })
       expect(wrapper.element).toMatchSnapshot()
