+++ spec/utils/animations.spec.ts
@@ -27,12 +27,8 @@ import {
   findNextFrameWithKeyframe,
   findPrevFrameWithKeyframe,
   getAfterKeyframe,
-  isSameKeyframeStatus,
   getLastFrame,
-  getSamePropRangeFrameMapByBoneId,
-  getSamePropRangeFrameMap,
 } from '/@/utils/animations'
-import { getKeyframeBoneDefaultPropsMap } from '/@/utils/keyframes'
 
 describe('utils/animations.ts', () => {
   describe('sortKeyframeMap', () => {
@@ -72,25 +68,6 @@ describe('utils/animations.ts', () => {
           1
         )
       ).toEqual(getKeyframeBone({ frame: 3 }))
-    })
-  })
-
-  describe('isSameKeyframeStatus', () => {
-    it('return false if different transform', () => {
-      expect(
-        isSameKeyframeStatus(
-          getKeyframeBone(),
-          getKeyframeBone({ rotate: getKeyframePoint({ value: 10 }) })
-        )
-      ).toBe(false)
-    })
-    it('return true if same transform', () => {
-      expect(
-        isSameKeyframeStatus(
-          getKeyframeBone({ rotate: getKeyframePoint({ value: 10 }) }),
-          getKeyframeBone({ rotate: getKeyframePoint({ value: 10 }) })
-        )
-      ).toBe(true)
     })
   })
 
@@ -296,128 +273,6 @@ describe('utils/animations.ts', () => {
           5
         )
       ).toBe(5)
-    })
-  })
-
-  describe('getSamePropRangeFrameMapByBoneId', () => {
-    it('get same range frame map by bone id', () => {
-      const t1 = getKeyframePoint({ value: 0 })
-      const t2 = getKeyframePoint({ value: 45 })
-      const res = getSamePropRangeFrameMapByBoneId({
-        a: [
-          getKeyframeBone({ frame: 0, rotate: t1 }),
-          getKeyframeBone({ frame: 3, rotate: t1 }),
-          getKeyframeBone({ frame: 5, rotate: t2 }),
-        ],
-        b: [getKeyframeBone({ frame: 0, rotate: t1 })],
-        c: [],
-      })
-      expect(res.a).toEqual({
-        0: { all: 0, ...getKeyframeBoneDefaultPropsMap(() => 0), rotate: 3 },
-        3: { all: 0, ...getKeyframeBoneDefaultPropsMap(() => 0) },
-        5: { all: 0, ...getKeyframeBoneDefaultPropsMap(() => 0) },
-      })
-      expect(res.b).toEqual({
-        0: { all: 0, ...getKeyframeBoneDefaultPropsMap(() => 0) },
-      })
-      expect(res.c).toEqual({})
-    })
-  })
-
-  describe('getSamePropRangeFrameMap', () => {
-    const t1 = getKeyframePoint({ value: 0 })
-    const t2 = getKeyframePoint({ value: 45 })
-
-    it('translateX', () => {
-      const list = [
-        getKeyframeBone({ frame: 0, translateX: t1 }),
-        getKeyframeBone({ frame: 3, translateX: t1 }),
-        getKeyframeBone({ frame: 5, translateX: t1 }),
-        getKeyframeBone({ frame: 5, translateX: t2 }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0).translateX).toEqual(3)
-      expect(getSamePropRangeFrameMap(list, 1).translateX).toEqual(2)
-      expect(getSamePropRangeFrameMap(list, 2).translateX).toEqual(0)
-    })
-    it('translateY', () => {
-      const list = [
-        getKeyframeBone({ frame: 0, translateY: t1 }),
-        getKeyframeBone({ frame: 3, translateY: t1 }),
-        getKeyframeBone({ frame: 5, translateY: t1 }),
-        getKeyframeBone({ frame: 5, translateY: t2 }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0).translateY).toEqual(3)
-      expect(getSamePropRangeFrameMap(list, 1).translateY).toEqual(2)
-      expect(getSamePropRangeFrameMap(list, 2).translateY).toEqual(0)
-    })
-    it('rotate', () => {
-      const list = [
-        getKeyframeBone({ frame: 0, rotate: t1 }),
-        getKeyframeBone({ frame: 3, rotate: t1 }),
-        getKeyframeBone({ frame: 5, rotate: t1 }),
-        getKeyframeBone({ frame: 5, rotate: t2 }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0).rotate).toEqual(3)
-      expect(getSamePropRangeFrameMap(list, 1).rotate).toEqual(2)
-      expect(getSamePropRangeFrameMap(list, 2).rotate).toEqual(0)
-    })
-    it('scaleX', () => {
-      const list = [
-        getKeyframeBone({ frame: 0, scaleX: t1 }),
-        getKeyframeBone({ frame: 3, scaleX: t1 }),
-        getKeyframeBone({ frame: 5, scaleX: t1 }),
-        getKeyframeBone({ frame: 5, scaleX: t2 }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0).scaleX).toEqual(3)
-      expect(getSamePropRangeFrameMap(list, 1).scaleX).toEqual(2)
-      expect(getSamePropRangeFrameMap(list, 2).scaleX).toEqual(0)
-    })
-    it('scaleY', () => {
-      const list = [
-        getKeyframeBone({ frame: 0, scaleY: t1 }),
-        getKeyframeBone({ frame: 3, scaleY: t1 }),
-        getKeyframeBone({ frame: 5, scaleY: t1 }),
-        getKeyframeBone({ frame: 5, scaleY: t2 }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0).scaleY).toEqual(3)
-      expect(getSamePropRangeFrameMap(list, 1).scaleY).toEqual(2)
-      expect(getSamePropRangeFrameMap(list, 2).scaleY).toEqual(0)
-    })
-    it('all prop equals min value', () => {
-      const list = [
-        getKeyframeBone({
-          frame: 0,
-          translateX: t1,
-          translateY: t1,
-          rotate: t1,
-          scaleX: t1,
-          scaleY: t1,
-        }),
-        getKeyframeBone({
-          frame: 3,
-          translateX: t1,
-          translateY: t1,
-          rotate: t1,
-          scaleX: t1,
-          scaleY: t1,
-        }),
-        getKeyframeBone({
-          frame: 4,
-          translateY: t1,
-          rotate: t1,
-          scaleX: t1,
-          scaleY: t1,
-        }),
-      ]
-      expect(getSamePropRangeFrameMap(list, 0)).toEqual({
-        all: 3,
-        ...getKeyframeBoneDefaultPropsMap(() => 3),
-      })
-      expect(getSamePropRangeFrameMap(list, 1)).toEqual({
-        all: 0,
-        ...getKeyframeBoneDefaultPropsMap(() => 1),
-        translateX: 0,
-      })
     })
   })
