+++ spec/utils/keyframes/keyframeBone.spec.ts
@@ -19,7 +19,12 @@ Copyright (C) 2021, Tomoya Komiyama.
 
 import { getTransform } from '/@/models'
 import { getKeyframeBone, getKeyframePoint } from '/@/models/keyframe'
-import * as target from '/@/utils/keyframes'
+import * as target from '/@/utils/keyframes/keyframeBone'
+import {
+  getKeyframeDefaultPropsMap,
+  getSamePropRangeFrameMap,
+  getSamePropRangeFrameMapByBoneId,
+} from '/@/utils/keyframes/keyframeBone'
 
 describe('utils/keyframes/index.ts', () => {
   describe('interpolateKeyframe', () => {
@@ -314,6 +319,128 @@ describe('utils/keyframes/index.ts', () => {
       expect(ret.rotate).toEqual([key3])
       expect(ret.scaleX).toEqual([key3])
       expect(ret.scaleY).toEqual([key3])
+    })
+  })
+
+  describe('getSamePropRangeFrameMapByBoneId', () => {
+    it('get same range frame map by bone id', () => {
+      const t1 = getKeyframePoint({ value: 0 })
+      const t2 = getKeyframePoint({ value: 45 })
+      const res = getSamePropRangeFrameMapByBoneId({
+        a: [
+          getKeyframeBone({ frame: 0, rotate: t1 }),
+          getKeyframeBone({ frame: 3, rotate: t1 }),
+          getKeyframeBone({ frame: 5, rotate: t2 }),
+        ],
+        b: [getKeyframeBone({ frame: 0, rotate: t1 })],
+        c: [],
+      })
+      expect(res.a).toEqual({
+        0: { all: 0, ...getKeyframeDefaultPropsMap(() => 0), rotate: 3 },
+        3: { all: 0, ...getKeyframeDefaultPropsMap(() => 0) },
+        5: { all: 0, ...getKeyframeDefaultPropsMap(() => 0) },
+      })
+      expect(res.b).toEqual({
+        0: { all: 0, ...getKeyframeDefaultPropsMap(() => 0) },
+      })
+      expect(res.c).toEqual({})
+    })
+  })
+
+  describe('getSamePropRangeFrameMap', () => {
+    const t1 = getKeyframePoint({ value: 0 })
+    const t2 = getKeyframePoint({ value: 45 })
+
+    it('translateX', () => {
+      const list = [
+        getKeyframeBone({ frame: 0, translateX: t1 }),
+        getKeyframeBone({ frame: 3, translateX: t1 }),
+        getKeyframeBone({ frame: 5, translateX: t1 }),
+        getKeyframeBone({ frame: 5, translateX: t2 }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0).translateX).toEqual(3)
+      expect(getSamePropRangeFrameMap(list, 1).translateX).toEqual(2)
+      expect(getSamePropRangeFrameMap(list, 2).translateX).toEqual(0)
+    })
+    it('translateY', () => {
+      const list = [
+        getKeyframeBone({ frame: 0, translateY: t1 }),
+        getKeyframeBone({ frame: 3, translateY: t1 }),
+        getKeyframeBone({ frame: 5, translateY: t1 }),
+        getKeyframeBone({ frame: 5, translateY: t2 }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0).translateY).toEqual(3)
+      expect(getSamePropRangeFrameMap(list, 1).translateY).toEqual(2)
+      expect(getSamePropRangeFrameMap(list, 2).translateY).toEqual(0)
+    })
+    it('rotate', () => {
+      const list = [
+        getKeyframeBone({ frame: 0, rotate: t1 }),
+        getKeyframeBone({ frame: 3, rotate: t1 }),
+        getKeyframeBone({ frame: 5, rotate: t1 }),
+        getKeyframeBone({ frame: 5, rotate: t2 }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0).rotate).toEqual(3)
+      expect(getSamePropRangeFrameMap(list, 1).rotate).toEqual(2)
+      expect(getSamePropRangeFrameMap(list, 2).rotate).toEqual(0)
+    })
+    it('scaleX', () => {
+      const list = [
+        getKeyframeBone({ frame: 0, scaleX: t1 }),
+        getKeyframeBone({ frame: 3, scaleX: t1 }),
+        getKeyframeBone({ frame: 5, scaleX: t1 }),
+        getKeyframeBone({ frame: 5, scaleX: t2 }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0).scaleX).toEqual(3)
+      expect(getSamePropRangeFrameMap(list, 1).scaleX).toEqual(2)
+      expect(getSamePropRangeFrameMap(list, 2).scaleX).toEqual(0)
+    })
+    it('scaleY', () => {
+      const list = [
+        getKeyframeBone({ frame: 0, scaleY: t1 }),
+        getKeyframeBone({ frame: 3, scaleY: t1 }),
+        getKeyframeBone({ frame: 5, scaleY: t1 }),
+        getKeyframeBone({ frame: 5, scaleY: t2 }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0).scaleY).toEqual(3)
+      expect(getSamePropRangeFrameMap(list, 1).scaleY).toEqual(2)
+      expect(getSamePropRangeFrameMap(list, 2).scaleY).toEqual(0)
+    })
+    it('all prop equals min value', () => {
+      const list = [
+        getKeyframeBone({
+          frame: 0,
+          translateX: t1,
+          translateY: t1,
+          rotate: t1,
+          scaleX: t1,
+          scaleY: t1,
+        }),
+        getKeyframeBone({
+          frame: 3,
+          translateX: t1,
+          translateY: t1,
+          rotate: t1,
+          scaleX: t1,
+          scaleY: t1,
+        }),
+        getKeyframeBone({
+          frame: 4,
+          translateY: t1,
+          rotate: t1,
+          scaleX: t1,
+          scaleY: t1,
+        }),
+      ]
+      expect(getSamePropRangeFrameMap(list, 0)).toEqual({
+        all: 3,
+        ...getKeyframeDefaultPropsMap(() => 3),
+      })
+      expect(getSamePropRangeFrameMap(list, 1)).toEqual({
+        all: 0,
+        ...getKeyframeDefaultPropsMap(() => 1),
+        translateX: 0,
+      })
     })
   })
 })
